requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crispy-forms==1.8.1
django-filter==2.2.0
django-serializer==0.0.8
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.1.1
psycopg2==2.8.4
pytz==2019.3
Serializer==0.2.1
simplejson==3.17.0
sqlparse==0.3.0
whitenoise==5.0.1
